Shutdown channel correctly
It is currently crash the runtime in shutdown function, that's why
we have a lot of exceptionCaught warning logs. I suppose there is a
bug in Scala compiler, that causes this. The closue doesn't seem to
have the correct type in runtime. The fix is annotated the type to
guide compiler.