send fewer `crowd` websocket events in very large rooms
we have broadcasts with 10k+ members.
the most frequent message sent to all members is `crowd`, once per
second.
if the room contains more than 100 members,
we only send `crowd` 10% of the time.
if the room has 1000 members, we send 1% of the time.

importantly, if the named users of the crowd change, then we always
send the message. That's necessary to know which contributors and
streamers are present.