batch study versioned messages
to better handle exceptional broadcasts.

the big ws spikes we see are due to events where many moves are received from the source at once, and sent to all viewers immediately

say we suddenly get 100 new moves, and there are 10k viewers. Each viewer needs to receive 100 addNode messages to update their study state. That's a million messages.

What currently happens is that for each message, lila-ws sends it to all 10k viewers. Then it proceeds to the next message, and so on.

It sends

msg 1 to viewer 1
msg 1 to viewer 2
msg 1 to viewer 3
[...]
msg 1 to viewer 10000
msg 2 to viewer 1
msg 2 to viewer 2
[...]
msg 3 to viewer 1
[...]

I'm studying the possibility of adding message batching into lila-ws, so that it can instead do:

msg [1 to 100] to viewer 1
msg [1 to 100] to viewer 2
[...]

which results in 10k big messages to send, instead of 1M small messages